Move test ui into demo module
