+++ addon/assembly/MANIFEST.MF
