+++ addon/assembly/assembly.xml
