+++ addon/src/main/java/org/vaadin/jouni/animator/Animator.java
