+++ addon/src/main/java/org/vaadin/jouni/animator/AnimatorWidgetset.gwt.xml
