+++ addon/src/main/java/org/vaadin/jouni/animator/client/AnimatorClientRpc.java
