+++ demo/src/main/resources/org/vaadin/jouni/animator/DemoWidgetSet.gwt.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.5.1//EN" "http://google-web-toolkit.googlecode.com/svn/tags/2.5.1/distro-source/core/src/gwt-module.dtd">
+<module>
+    <!--
+        This file is automatically updated based on new dependencies by the
+        goal "vaadin:update-widgetset".
+    -->
+
+    <!-- Inherit add-on widgetset -->
+    <inherits name="org.vaadin.jouni.animator.AnimatorWidgetset" />
+    <inherits name="org.vaadin.jouni.dom.DomWidgetset" />
+
+    <!--
+     Uncomment the following to compile the widgetset for one browser only.
+     Multiple browsers can be specified as a comma separated list. The
+     supported user agents at the moment of writing were:
+     ie8,ie9,gecko1_8,safari,opera
+     The value gecko1_8 is used for Firefox and safari is used for webkit
+     based browsers including Google Chrome.
+    -->
+    <!-- <set-property name="user.agent" value="safari"/> -->
+
+    <inherits name="com.vaadin.DefaultWidgetSet" />
+
+</module>
